+++ src/api.ts
@@ -8,13 +8,10 @@ import AADTokenInfo from "./auth/AADTokenInfo";
 import * as mongoose from 'mongoose';
 import {json as jsonParser} from 'express';
 import EventSchema from "./schema/Event";
-import {GraphAPI} from "./auth/GraphAPI";
 const APIRouter = Router();
-const membersRouter = Router();
 
 mongoose.connect(process.env.MONGO_URI!);
 const db = mongoose.connection;
-const graphAPI = new GraphAPI(process.env.GRAPH_API_CLIENT!, process.env.GRAPH_API_SECRET!);
 
 db.on("error", console.error.bind(console, "connection error:"));
 
@@ -101,9 +98,6 @@ APIRouter.patch('/private/events/:id', (req: Request, res: Response) => {
     })
 })
 
-membersRouter.get('/checkStatus/:netID', async (req: Request, res: Response) => {
-    var memberStatus = await graphAPI.isPaidMember(req.params.netID);
-    return res.send(memberStatus);
-})
 
-export {APIRouter, membersRouter};
+
+export {APIRouter};
